This PR converts all of our stdlib JLLs to be "lazy" JLLs, which provides a nice improvement to startup latency, and reduces overall memory footprint by not loading libraries that are never called.

Example timings:

```
# current master
$ hyperfine -w 1 -r 20 ./julia
Benchmark 1: ./julia
  Time (mean ± σ):     152.2 ms ±   1.8 ms    [User: 272.5 ms, System: 86.7 ms]
  Range (min … max):   149.3 ms … 155.3 ms    20 runs
```

```
# this PR
$ hyperfine -w 1 -r 20 ./julia
Benchmark 1: ./julia
  Time (mean ± σ):     122.4 ms ±   2.3 ms    [User: 86.7 ms, System: 33.2 ms]
  Range (min … max):   119.6 ms … 130.4 ms    20 runs
```

Depends on:
- [x] https://github.com/JuliaLang/julia/pull/50074
- [ ] https://github.com/JuliaLang/Downloads.jl/pull/229