@staticfloat 2023-07-27
> what's the rss difference?

Not much, 2 MB out of 190 MB.  I'm a little surprised by that, but I think it's likely because most of the pages we're eliminating here are shared.