@PallHaraldsson 2023-07-28
>converts all of our stdlib JLLs to be "lazy" JLLs, which provides a nice improvement to startup latency

This is for stdlib JLLs only, right? Still great, but the idea would translate to all JLLs? Easy to extend that way, and might also make for nice speedup? Each one is seemingly usually very cheap, but sometimes many used, so add up. I was also (alternatively) thinking parallel loading of JLLs, seems doable, probably easier than for non-JLL.