@staticfloat 2023-07-28
> This is for stdlib JLLs only, right? Still great, but the idea would translate to all JLLs?

Yes, that's the idea, but we need more information than exists in JLLs right now; we need to encode the dependency graph of each library on every other library.

> Is this too large to backport? Not right away, but ion time for 1.10?

I don't think this is a huge feature, and my time to work on it post-JuliaCon will be limited, so I would expect it in 1.11.